feat: add invalid input operation
