+++ package.json
@@ -2,10 +2,10 @@
   "name": "file-manager",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "greeting.js",
   "type": "module",
   "scripts": {
-    "start": "node src/index.js"
+    "start": "node src/greeting.js"
   },
   "author": "",
   "license": "ISC"
